transactions/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-decouple==3.8
Pillow==11.0.0
django-allauth==65.3.0
python-dateutil==2.8.2
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.2.0
